@@ -18,44 +18,44 @@
     "lint": "eslint --cache --ext .js,.jsx,.tsx,.ts ./src"
   },
   "dependencies": {
-    "@ant-design/icons": "^4.7.0",
-    "@reduxjs/toolkit": "^1.8.4",
-    "@toast-ui/editor": "^3.2.0",
-    "antd": "^4.22.4",
-    "axios": "^0.27.2",
+    "@ant-design/icons": "^4.8.0",
+    "@reduxjs/toolkit": "^1.9.1",
+    "@toast-ui/editor": "^3.2.1",
+    "antd": "^5.1.0",
+    "axios": "^1.2.1",
     "blueimp-md5": "^2.19.0",
     "bytes": "^3.1.2",
-    "classnames": "^2.3.1",
+    "classnames": "^2.3.2",
+    "dayjs": "^1.11.7",
     "lodash": "^4.17.21",
-    "moment": "^2.29.4",
-    "query-string": "^7.1.1",
+    "query-string": "^8.1.0",
     "react": "^18.2.0",
-    "react-beautiful-dnd": "^13.1.0",
+    "react-beautiful-dnd": "^13.1.1",
     "react-dom": "^18.2.0",
-    "react-redux": "^8.0.2",
-    "react-router-dom": "^6.3.0",
-    "recharts": "^2.1.13",
+    "react-redux": "^8.0.5",
+    "react-router-dom": "^6.5.0",
+    "recharts": "^2.2.0",
     "redux": "^4.2.0",
-    "redux-thunk": "^2.4.1",
-    "sass": "^1.54.4",
+    "redux-thunk": "^2.4.2",
+    "sass": "^1.57.1",
     "use-keep-state": "^1.3.0"
   },
   "devDependencies": {
     "@types/blueimp-md5": "^2.18.0",
     "@types/bytes": "^3.1.1",
-    "@types/lodash": "^4.14.182",
-    "@types/node": "^18.7.2",
-    "@types/react": "^18.0.17",
-    "@types/react-beautiful-dnd": "^13.1.2",
-    "@types/react-dom": "^18.0.6",
+    "@types/lodash": "^4.14.191",
+    "@types/node": "^18.11.17",
+    "@types/react": "^18.0.26",
+    "@types/react-beautiful-dnd": "^13.1.3",
+    "@types/react-dom": "^18.0.9",
     "@types/react-redux": "^7.1.24",
     "@types/react-router-dom": "^5.3.3",
-    "@typescript-eslint/eslint-plugin": "^5.33.0",
-    "@typescript-eslint/parser": "^5.33.0",
+    "@typescript-eslint/eslint-plugin": "^5.47.0",
+    "@typescript-eslint/parser": "^5.47.0",
     "@vitejs/plugin-react-refresh": "^1.3.6",
-    "eslint": "^8.21.0",
-    "eslint-plugin-react": "7.30.1",
-    "typescript": "^4.7.4",
-    "vite": "^3.0.7"
+    "eslint": "^8.30.0",
+    "eslint-plugin-react": "7.31.11",
+    "typescript": "^4.9.4",
+    "vite": "^4.0.2"
   }
 }
